requirements.txt
asgiref==3.8.1
attrs==25.3.0
cffi==1.17.1
cryptography==44.0.2
Django==4.2.21
django-cors-headers==4.7.0
djangorestframework==3.16.0
drf-spectacular==0.28.0
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
mysqlclient==2.2.7
pillow==11.2.1
pycparser==2.22
PyMySQL==1.1.1
PyYAML==6.0.2
referencing==0.36.2
rpds-py==0.24.0
sqlparse==0.5.3
uritemplate==4.1.1
djangorestframework-simplejwt>=5.5.0
